requirements.txt
beautifulsoup4==4.11.2
feedparser==6.0.10
pandas==1.5.3
requests==2.28.2
nltk==3.8.1
newspaper3k==0.2.8
python-dateutil==2.8.2
Unidecode==1.3.6
lxml_html_clean
urllib3==1.26.14
